// frontend_native/features/shared/api/generated/model/signinRequest.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * Awesome Template API
 * Awesome Template バックエンドサービスの API ドキュメントです。
 * OpenAPI spec version: 1.0.0
 */

/**
 * ユーザーがサインインするためのリクエストペイロードです
 */
export interface SigninRequest {
  /** ユーザーのメールアドレス */
  email: string;
  /** ユーザーのパスワード */
  password: string;
}